[ibb] Use UInt16 for 'seq' and fix its handling
Fixes a off-by-one error when incrementing 'seq'. Thanks to Kim
Alvefur <zash@zash.se> for spotting this.

+++ smack-extensions/src/test/java/org/jivesoftware/smackx/bytestreams/ibb/packet/DataPacketExtensionTest.java
@@ -75,7 +75,7 @@ public class DataPacketExtensionTest extends SmackTestSuite {
     public void shouldSetAllFieldsCorrectly() {
         DataPacketExtension data = new DataPacketExtension("sessionID", 0, "data");
         assertEquals("sessionID", data.getSessionID());
-        assertEquals(0, data.getSeq());
+        assertEquals(0, data.getSeq().intValue());
         assertEquals("data", data.getData());
     }
